// src/app/modules/check-in-summary/src/lib/enums/translations-keys.enum.ts
export enum TranslationKeys {
  // General
  CheckIn_CheckInNotAvailable = 'CheckIn.CheckInNotAvailable',

  // Passenger Status
  CheckIn_NotAllowed = 'CheckIn.NotAllowed',

  // Seat Assignment
  CheckIn_SeatAssigned_None = 'CheckIn.SeatAssigned_None',

  // Select All Passengers
  CheckIn_SelectAllPassengers = 'CheckIn.SelectAllPassengers',
  CheckIn_A11y_Passengers_SelectAll = 'CheckIn.A11y.Passengers_SelectAll',
  CheckIn_A11y_Passengers_DeselectAll = 'CheckIn.A11y.Passengers_DeselectAll',

  // Airport Counter
  CheckIn_AirportCounter_SeatAvailability_Info = 'CheckIn.AirportCounter.SeatAvailability_Info',
  CheckIn_AirportCounter_SeatAvailability_Alert = 'CheckIn.AirportCounter.SeatAvailability_Alert',

  // Error Messages
  CheckIn_Modal_Error_Title = 'CheckIn.Modal.Error_Title',
  CheckIn_Modal_AtLeastOnePassengerRequired_Message = 'CheckIn.Modal.AtLeastOnePassengerRequired_Message',

  // Journey Region
  CheckIn_JourneyRegion_Checked_None = 'CheckIn.JourneyRegion.Checked_None',
  CheckIn_JourneyRegion_Checked_All = 'CheckIn.JourneyRegion.Checked_All',
  CheckIn_JourneyRegion_Checked_One = 'CheckIn.JourneyRegion.Checked_One',
  CheckIn_JourneyRegion_Checked_Some = 'CheckIn.JourneyRegion.Checked_Some',
  CheckIn_JourneyRegion_AriaLabel = 'CheckIn.JourneyRegion.AriaLabel',
}